Fix 2 errors: separate classes and interfaces from 1 class, creation and checking endure to new methods
