+++ src/com/homework6/MailMessage.java
@@ -1,0 +1,38 @@
+package com.homework6;
+
+/**
+ * Class for saving information about mail message.
+ */
+public class MailMessage implements Sendable<String> {
+  private String from;
+  private String to;
+  private String content;
+
+  public MailMessage(String from, String to, String content) {
+    this.from = from;
+    this.to = to;
+    this.content = content;
+  }
+
+  public String getFrom() {
+    return from;
+  }
+
+  public String getTo() {
+    return to;
+  }
+
+  public String getContent() {
+    return content;
+  }
+
+  @Override
+  public String getKey() {
+    return getTo();
+  }
+
+  @Override
+  public String getValue() {
+    return getContent();
+  }
+}
