+++ src/com/homework6/Sendable.java
@@ -1,0 +1,10 @@
+package com.homework6;
+
+/**
+ * Interface for elements, which work with consumer.
+ */
+public interface Sendable<T> {
+  String getKey();
+
+  T getValue();
+}
