+++ src/com/homework6/StreamApiTaskTests.java
@@ -2,14 +2,11 @@ package com.homework6;
 
 import java.util.Arrays;
 import java.util.Collections;
-import java.util.HashMap;
-import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
-import java.util.function.Consumer;
 
 /**
- * Solution for task with *: do work code.
+ * Solution for task with *: do code to compile and work.
  */
 public class StreamApiTaskTests {
   /**
@@ -17,6 +14,23 @@ public class StreamApiTaskTests {
    * @param args arguments to command line
    */
   public static void main(String[] args) {
+    List<MailMessage> messages = formMailMessages();
+    MailService<String> mailService = new MailService<>();
+    messages.stream().forEachOrdered(mailService);
+    Map<String, List<String>> mailBox = mailService.getMailBox();
+    String randomFrom = "...";
+    String randomTo = "...";
+    int randomSalary = 100;
+    checkMailBox(mailBox, randomTo);
+
+    List<Salary> salariesList = formSalaries(randomFrom, randomTo, randomSalary);
+    MailService<Integer> salaryService = new MailService<>();
+    salariesList.forEach(salaryService);
+    Map<String, List<Integer>> salaries = salaryService.getMailBox();
+    checkSalaries(salaries, salariesList, randomTo, randomSalary);
+  }
+
+  private static List<MailMessage> formMailMessages() {
     MailMessage firstMessage = new MailMessage(
         "Robert Howard",
         "H.P. Lovecraft",
@@ -40,16 +54,10 @@ public class StreamApiTaskTests {
         "Я так и не понял Интерстеллар."
     );
 
-    List<MailMessage> messages = Arrays.asList(
-        firstMessage, secondMessage, thirdMessage
-    );
+    return Arrays.asList(firstMessage, secondMessage, thirdMessage);
+  }
 
-    MailService<String> mailService = new MailService<>();
-
-    messages.stream().forEachOrdered(mailService);
-
-    Map<String, List<String>> mailBox = mailService.getMailBox();
-
+  private static void checkMailBox(Map<String, List<String>> mailBox, String randomTo) {
     assert mailBox.get("H.P. Lovecraft").equals(
         Arrays.asList(
             "This \"The Shadow over Innsmouth\" story is real masterpiece, Howard!"
@@ -64,121 +72,24 @@ public class StreamApiTaskTests {
         )
     ) : "wrong mailService mailbox content (2)";
 
-    String randomFrom = "...";
-    String randomTo = "...";
-    int randomSalary = 100;
-
     assert mailBox.get(randomTo)
         .equals(Collections.<String>emptyList()) : "wrong mailService mailbox content (3)";
+  }
 
+  private static List<Salary> formSalaries(String from, String to, int salary) {
     Salary salary1 = new Salary("Facebook", "Mark Zuckerberg", 1);
     Salary salary2 = new Salary("FC Barcelona", "Lionel Messi", Integer.MAX_VALUE);
-    Salary salary3 = new Salary(randomFrom, randomTo, randomSalary);
+    Salary salary3 = new Salary(from, to, salary);
+    return Arrays.asList(salary1, salary2, salary3);
+  }
 
-    MailService<Integer> salaryService = new MailService<>();
-
-    Arrays.asList(salary1, salary2, salary3).forEach(salaryService);
-
-    Map<String, List<Integer>> salaries = salaryService.getMailBox();
-    assert salaries.get(salary1.getTo())
+  private static void checkSalaries(Map<String, List<Integer>> salaries, List<Salary> salariesList,
+      String randomTo, int randomSalary) {
+    assert salaries.get(salariesList.get(0).getTo())
         .equals(Arrays.asList(1)) : "wrong salaries mailbox content (1)";
-    assert salaries.get(salary2.getTo())
+    assert salaries.get(salariesList.get(0).getTo())
         .equals(Arrays.asList(Integer.MAX_VALUE)) : "wrong salaries mailbox content (2)";
     assert salaries.get(randomTo)
         .equals(Arrays.asList(randomSalary)) : "wrong salaries mailbox content (3)";
-  }
-
-  public static class MailMessage implements Sendable<String> {
-    private String from;
-    private String to;
-    private String content;
-
-    public MailMessage(String from, String to, String content) {
-      this.from = from;
-      this.to = to;
-      this.content = content;
-    }
-
-    public String getFrom() {
-      return from;
-    }
-
-    public String getTo() {
-      return to;
-    }
-
-    public String getContent() {
-      return content;
-    }
-
-    @Override
-    public String getKey() {
-      return getTo();
-    }
-
-    @Override
-    public String getValue() {
-      return getContent();
-    }
-  }
-
-  public static class Salary implements Sendable<Integer> {
-    private String placeWork;
-    private String to;
-    private int salary;
-
-    public Salary(String placeWork, String to, int salary) {
-      this.placeWork = placeWork;
-      this.to = to;
-      this.salary = salary;
-    }
-
-    public String getTo() {
-      return to;
-    }
-
-    public int getSalary() {
-      return salary;
-    }
-
-    @Override
-    public String getKey() {
-      return getTo();
-    }
-
-    @Override
-    public Integer getValue() {
-      return getSalary();
-    }
-  }
-
-  public static class MailService<T> implements Consumer<Sendable<T>> {
-    private Map<String, List<T>> result = new MyHashMap<>();
-
-    @Override
-    public void accept(Sendable<T> sendableT) {
-      List<T> elements =
-          result.get(sendableT.getKey()) == Collections.EMPTY_LIST ? new LinkedList<T>() :
-              result.get(sendableT.getKey());
-      elements.add(sendableT.getValue());
-      result.put(sendableT.getKey(), elements);
-    }
-
-    public Map<String, List<T>> getMailBox() {
-      return result;
-    }
-  }
-
-  public static interface Sendable<T> {
-    String getKey();
-
-    T getValue();
-  }
-
-  public static class MyHashMap<K, V> extends HashMap<K, V> {
-    @Override
-    public V get(Object key) {
-      return super.get(key) == null ? (V) Collections.EMPTY_LIST : super.get(key);
-    }
   }
 }
